Multiple changes in the Java classes.
-Added the methods computeHashToHashesSimilarities (computes the
similarities between a hash and a map of hashes and returns them as a
map) and computeAllHashesSimilarities (computes the similarities between
all the hashes in a map and returns them as a map) to the class
UniformFuzzyHashes.java.

-Added the methods sortSimilarities (sorts a map of similarities between
a hash and a map of hashes by a type of similarity) and sortMap (sorts
any map by the same order as other sorted map).

-Hash statistics and characteristics removed due to underusage.

-Similarity cache removed from the class UniformFuzzyHash.java,
it is not necessary anymore since from now on it can be managed
outside the class thanks to the new methods
computeHashToHashesSimilarities and computeAllHashesSimilarities.

-SimilarityTypes enum moved to the class UniformFuzzyHash.java from the
class UniformFuzzyHashes.java. Added the methods similarity(hash,
similarityType) (computes and returns a type of similarity between two
hashes) and similarities (computes all the types of similarities between
two hashes and returns them as a map) to the class
UniformFuzzyHash.java.

-All the methods using a map of names (string) -> objects in the class
UniformFuzzyHashes.java changed to a generic map of identifiers (any
type) -> objects.

-Methods printHashToHashesSimilaritiesTable,
saveHashToHashesSimilaritiesAsCsv, printAllHashesSimilaritiesTable and
saveAllHashesSimilaritiesAsCsv in the class UniformFuzzyHashes.java
changed to receive a map of similarities.

-Added the methods collectionToMap (builds a map of identifiers ->
objects from a collection of objects, identifying them by index),
mapValuesToList (builds a list from the values in a map) and
mapKeysToList (builds a list from the keys in a map) to the class
UniformFuzzyHashes.java.

-Removed the methods using collections instead of maps in the class
UniformFuzzyHashes.java, they are not necessary anymore thanks to the
new methods collectionToMap, mapValuesToList, mapKeysToList and sortMap.